requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0


